+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Version 2.6.0
+
+### Added:
+
+1. Encryption for username & password in app_data_
+
+### Fixed/Changes:
+
+1. Improved github repo structure
+2. Fixed autocompleter insertion
+
 # Version 2.5.0
 
 ### Major updates